+++ frontends/cuse/cuse_device.h
@@ -32,12 +32,13 @@ extern "C"
 {
 #endif
 
+typedef ssize_t (*get_func_t)(uint8_t *buf, size_t buflen, void *int_data);
+
 void esdm_cuse_init_done(void *userdata);
 void esdm_cuse_open(fuse_req_t req, struct fuse_file_info *fi);
 void esdm_cuse_read_internal(fuse_req_t req, size_t size, off_t off,
 			     struct fuse_file_info *fi,
-			     ssize_t (*get)(uint8_t *buf, size_t buflen),
-			     int fallback_fd);
+			     get_func_t get, int fallback_fd);
 void esdm_cuse_write_internal(fuse_req_t req, const char *buf, size_t size,
 			      off_t off, struct fuse_file_info *fi,
 			      int fallback_fd);