+++ service-rpc/client/esdm_rpc_get_poolsize_c.c
@@ -24,6 +24,7 @@
 #include "esdm_rpc_service.h"
 #include "helper.h"
 #include "logger.h"
+#include "ptr_err.h"
 #include "ret_checkers.h"
 #include "visibility.h"
 
@@ -39,10 +40,10 @@ esdm_rpcc_get_poolsize_cb(const GetPoolsizeResponse *response,
 	struct esdm_poolsize_buf *buffer =
 				(struct esdm_poolsize_buf *)closure_data;
 
-	if (!response) {
+	if (IS_ERR(response)) {
 		logger(LOGGER_DEBUG, LOGGER_C_RPC,
 		       "missing data - connection interrupted\n");
-		buffer->ret = -EINTR;
+		buffer->ret = (int)PTR_ERR(response);
 		return;
 	}
 
@@ -51,14 +52,14 @@ esdm_rpcc_get_poolsize_cb(const GetPoolsizeResponse *response,
 }
 
 DSO_PUBLIC
-int esdm_rpcc_get_poolsize(unsigned int *poolsize)
+int esdm_rpcc_get_poolsize_int(unsigned int *poolsize, void *int_data)
 {
 	GetPoolsizeRequest msg = GET_POOLSIZE_REQUEST__INIT;
 	struct esdm_rpc_client_connection *rpc_conn;
 	struct esdm_poolsize_buf buffer;
 	int ret = 0;
 
-	CKINT(esdm_rpcc_get_unpriv_service(&rpc_conn));
+	CKINT(esdm_rpcc_get_unpriv_service(&rpc_conn, int_data));
 
 	buffer.ret = -ETIMEDOUT;
 
@@ -71,4 +72,10 @@ int esdm_rpcc_get_poolsize(unsigned int *poolsize)
 
 out:
 	return ret;
+}
+
+DSO_PUBLIC
+int esdm_rpcc_get_poolsize(unsigned int *poolsize)
+{
+	return esdm_rpcc_get_poolsize_int(poolsize, NULL);
 }