+++ service-rpc/client/esdm_rpc_get_write_wakeup_thresh_c.c
@@ -24,6 +24,7 @@
 #include "esdm_rpc_service.h"
 #include "helper.h"
 #include "logger.h"
+#include "ptr_err.h"
 #include "ret_checkers.h"
 #include "visibility.h"
 
@@ -38,10 +39,10 @@ static void esdm_rpcc_get_write_wakeup_thresh_cb(
 	struct esdm_write_wakeup_thresh_buf *buffer =
 			(struct esdm_write_wakeup_thresh_buf *)closure_data;
 
-	if (!response) {
+	if (IS_ERR(response)) {
 		logger(LOGGER_DEBUG, LOGGER_C_RPC,
 		       "missing data - connection interrupted\n");
-		buffer->ret = -EINTR;
+		buffer->ret = (int)PTR_ERR(response);
 		return;
 	}
 
@@ -50,14 +51,15 @@ static void esdm_rpcc_get_write_wakeup_thresh_cb(
 }
 
 DSO_PUBLIC
-int esdm_rpcc_get_write_wakeup_thresh(unsigned int *write_wakeup_thresh)
+int esdm_rpcc_get_write_wakeup_thresh_int(unsigned int *write_wakeup_thresh,
+					  void *int_data)
 {
 	GetWriteWakeupThreshRequest msg = GET_WRITE_WAKEUP_THRESH_REQUEST__INIT;
 	struct esdm_rpc_client_connection *rpc_conn;
 	struct esdm_write_wakeup_thresh_buf buffer;
 	int ret = 0;
 
-	CKINT(esdm_rpcc_get_unpriv_service(&rpc_conn));
+	CKINT(esdm_rpcc_get_unpriv_service(&rpc_conn, int_data));
 
 	buffer.ret = -ETIMEDOUT;
 
@@ -70,4 +72,10 @@ int esdm_rpcc_get_write_wakeup_thresh(unsigned int *write_wakeup_thresh)
 
 out:
 	return ret;
+}
+
+DSO_PUBLIC
+int esdm_rpcc_get_write_wakeup_thresh(unsigned int *write_wakeup_thresh)
+{
+	return esdm_rpcc_get_write_wakeup_thresh_int(write_wakeup_thresh, NULL);
 }