+++ service-rpc/client/esdm_rpc_rnd_add_entropy_c.c
@@ -24,6 +24,7 @@
 #include "esdm_rpc_service.h"
 #include "helper.h"
 #include "logger.h"
+#include "ptr_err.h"
 #include "ret_checkers.h"
 #include "visibility.h"
 
@@ -38,10 +39,10 @@ esdm_rpcc_rnd_add_entropy_cb(const RndAddEntropyResponse *response,
 	struct esdm_rnd_add_entropy_buf *buffer =
 			(struct esdm_rnd_add_entropy_buf *)closure_data;
 
-	if (!response) {
+	if (IS_ERR(response)) {
 		logger(LOGGER_DEBUG, LOGGER_C_RPC,
 		       "missing data - connection interrupted\n");
-		buffer->ret = -EINTR;
+		buffer->ret = (int)PTR_ERR(response);
 		return;
 	}
 
@@ -49,16 +50,17 @@ esdm_rpcc_rnd_add_entropy_cb(const RndAddEntropyResponse *response,
 }
 
 DSO_PUBLIC
-int esdm_rpcc_rnd_add_entropy(const uint8_t *entropy_buf,
-			      size_t entropy_buf_len,
-			      uint32_t entropy_cnt)
+int esdm_rpcc_rnd_add_entropy_int(const uint8_t *entropy_buf,
+				  size_t entropy_buf_len,
+				  uint32_t entropy_cnt,
+				  void *int_data)
 {
 	RndAddEntropyRequest msg = RND_ADD_ENTROPY_REQUEST__INIT;
 	struct esdm_rpc_client_connection *rpc_conn;
 	struct esdm_rnd_add_entropy_buf buffer;
 	int ret = 0;
 
-	CKINT(esdm_rpcc_get_priv_service(&rpc_conn));
+	CKINT(esdm_rpcc_get_priv_service(&rpc_conn, int_data));
 
 	buffer.ret = -ETIMEDOUT;
 
@@ -73,4 +75,13 @@ int esdm_rpcc_rnd_add_entropy(const uint8_t *entropy_buf,
 
 out:
 	return ret;
+}
+
+DSO_PUBLIC
+int esdm_rpcc_rnd_add_entropy(const uint8_t *entropy_buf,
+			      size_t entropy_buf_len,
+			      uint32_t entropy_cnt)
+{
+	return esdm_rpcc_rnd_add_entropy_int(entropy_buf, entropy_buf_len,
+					     entropy_cnt, NULL);
 }