+++ service-rpc/client/esdm_rpc_set_min_reseed_secs_c.c
@@ -23,6 +23,7 @@
 #include "esdm_rpc_service.h"
 #include "helper.h"
 #include "logger.h"
+#include "ptr_err.h"
 #include "ret_checkers.h"
 #include "visibility.h"
 
@@ -36,10 +37,10 @@ static void esdm_rpcc_set_min_reseed_secs_cb(
 	struct esdm_set_min_reseed_secs_buf *buffer =
 			(struct esdm_set_min_reseed_secs_buf *)closure_data;
 
-	if (!response) {
+	if (IS_ERR(response)) {
 		logger(LOGGER_DEBUG, LOGGER_C_RPC,
 		       "missing data - connection interrupted\n");
-		buffer->ret = -EINTR;
+		buffer->ret = (int)PTR_ERR(response);
 		return;
 	}
 
@@ -47,14 +48,14 @@ static void esdm_rpcc_set_min_reseed_secs_cb(
 }
 
 DSO_PUBLIC
-int esdm_rpcc_set_min_reseed_secs(unsigned int seconds)
+int esdm_rpcc_set_min_reseed_secs_int(unsigned int seconds, void *int_data)
 {
 	SetMinReseedSecsRequest msg = SET_MIN_RESEED_SECS_REQUEST__INIT;
 	struct esdm_rpc_client_connection *rpc_conn;
 	struct esdm_set_min_reseed_secs_buf buffer;
 	int ret = 0;
 
-	CKINT(esdm_rpcc_get_priv_service(&rpc_conn));
+	CKINT(esdm_rpcc_get_priv_service(&rpc_conn, int_data));
 
 	buffer.ret = -ETIMEDOUT;
 
@@ -66,4 +67,10 @@ int esdm_rpcc_set_min_reseed_secs(unsigned int seconds)
 
 out:
 	return ret;
+}
+
+DSO_PUBLIC
+int esdm_rpcc_set_min_reseed_secs(unsigned int seconds)
+{
+	return esdm_rpcc_set_min_reseed_secs_int(seconds, NULL);
 }